MessagingSummaryItem css in js migration
